added rectangular raster support added cov trasnform

+++ KernelRasterTransform/IRasterTransform.cs
@@ -2,6 +2,6 @@
 {
     public interface IRasterTransform
     {
-        float Transform(IEnumerable<float> values);
+        double Transform(IEnumerable<double> values);
     }
 }
